Rename `RlDropdownContainer` component to `Dropdown`
