+++ app/components/dropdown.hbs
@@ -1,0 +1,4 @@
+{{yield (hash
+  Trigger=(component "dropdown/trigger")
+  Content=(component "dropdown/content")
+)}}
