+++ app/components/dropdown/trigger.js
@@ -1,7 +1,7 @@
 import Component from '@ember/component';
 import { computed } from '@ember/object';
 
-import RlDropdownContainer from './rl-dropdown-container';
+import Dropdown from '../dropdown';
 
 export default Component.extend({
   classNames: ['rl-dropdown-toggle'],
@@ -19,7 +19,7 @@ export default Component.extend({
   }),
 
   dropdownContainer: computed(function() {
-    return this.nearestOfType(RlDropdownContainer);
+    return this.nearestOfType(Dropdown);
   }),
 
   disabled: false,
