--- app/components/rl-dropdown-container.hbs
@@ -1,4 +1,0 @@
-{{yield (hash
-  Trigger=(component "rl-dropdown-toggle")
-  Content=(component "rl-dropdown")
-)}}
