+++ app/templates/keyword/index.hbs
@@ -17,7 +17,7 @@
 
   <div class='sort' data-test-keyword-sort>
     <span class='small'>Sort by</span>
-    <RlDropdownContainer class="dropdown-container" as |dd|>
+    <Dropdown class="dropdown-container" as |dd|>
       <dd.Trigger @tagName="a" class="dropdown" data-test-current-order>
         {{svg-jar "sort"}}
         {{ this.currentSortBy }}
@@ -46,7 +46,7 @@
           </LinkTo>
         </li>
       </dd.Content>
-    </RlDropdownContainer>
+    </Dropdown>
   </div>
 </div>
