+++ app/templates/me/crates.hbs
@@ -18,7 +18,7 @@
 
   <div class='sort'>
     <span class='small'>Sort by</span>
-    <RlDropdownContainer class="dropdown-container" as |dd|>
+    <Dropdown class="dropdown-container" as |dd|>
       <dd.Trigger @tagName="a" class="dropdown">
         {{svg-jar "sort"}}
         {{ this.currentSortBy }}
@@ -47,7 +47,7 @@
           </LinkTo>
         </li>
       </dd.Content>
-    </RlDropdownContainer>
+    </Dropdown>
   </div>
 </div>
