+++ app/templates/me/following.hbs
@@ -16,7 +16,7 @@
 
   <div class='sort'>
     <span class='small'>Sort by</span>
-    <RlDropdownContainer class="dropdown-container" as |dd|>
+    <Dropdown class="dropdown-container" as |dd|>
       <dd.Trigger @tagName="a" class="dropdown">
         {{svg-jar "sort"}}
         {{ this.currentSortBy }}
@@ -35,7 +35,7 @@
           </LinkTo>
         </li>
       </dd.Content>
-    </RlDropdownContainer>
+    </Dropdown>
   </div>
 </div>
