+++ app/templates/search.hbs
@@ -29,7 +29,7 @@
     <div class='sort' data-test-search-sort>
       <span class='small'>Sort by </span>
 
-      <RlDropdownContainer class="dropdown-container" as |dd|>
+      <Dropdown class="dropdown-container" as |dd|>
         <dd.Trigger @tagName="a" class="dropdown" data-test-current-order>
           {{svg-jar "sort"}}
           {{ this.currentSortBy }}
@@ -58,7 +58,7 @@
             </LinkTo>
           </li>
         </dd.Content>
-      </RlDropdownContainer>
+      </Dropdown>
     </div>
   </div>
